parameter dictionary example in mlp-gluon.ipynb
After having been through all of the documents, I believe showing access to the parameter dictionary at this point in the tutorial would be a great aid to most Gluon beginners.  It will also help be able to peak under the hood of the upcoming section: Plumbing: A look under the hood of gluon.